requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
openai>=1.20.0
pypdf>=4.0.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=1.1.0
lxml>=5.0.0
